requirements.txt
django==5.2.7
djangorestframework==3.15.2
drf-spectacular==0.27.2
environs==11.0.0
django-cors-headers==4.6.0
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
dj-database-url==2.1.0
whitenoise==6.6.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
stripe==10.0.0